+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added config file option to change shell used to execute commands.
 - Added sort options for interactive view. Defaults to sort by label.
 - Added command line switch to change command database file.
+- Added support for editing really long lines (scrolling edit).
 - Fixed missing cursor when editing on some terminals.
 
 ## [0.7.0] - 2021-02-09